requirements.txt
# Backend dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
polars>=1.7.0
pyarrow>=21.0.0
duckdb>=1.4.0
numpy>=2.0.0
scipy>=1.12.0
scikit-learn>=1.5.0
pyyaml>=6.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
